+++ saleor/warehouse/tests/webhooks/subscriptions/test_channel_stock_events.py
@@ -1,10 +1,93 @@
 import json
 
 import graphene
+import pytest
 
-from ....warehouse.interface import VariantChannelStockInfo
-from ...event_types import WebhookEventAsyncType
-from ...transport.asynchronous.transport import create_deliveries_for_subscriptions
+from .....webhook.event_types import WebhookEventAsyncType
+from .....webhook.transport.asynchronous.transport import (
+    create_deliveries_for_subscriptions,
+)
+from ....interface import VariantChannelStockInfo
+
+PRODUCT_VARIANT_OUT_OF_STOCK_IN_CHANNEL = """
+    subscription{
+      event{
+        ...on ProductVariantOutOfStockInChannel{
+          productVariant{ id }
+          channel{ slug }
+        }
+      }
+    }
+"""
+
+PRODUCT_VARIANT_BACK_IN_STOCK_IN_CHANNEL = """
+    subscription{
+      event{
+        ...on ProductVariantBackInStockInChannel{
+          productVariant{ id }
+          channel{ slug }
+        }
+      }
+    }
+"""
+
+PRODUCT_VARIANT_OUT_OF_STOCK_FOR_CLICK_AND_COLLECT = """
+    subscription{
+      event{
+        ...on ProductVariantOutOfStockForClickAndCollect{
+          productVariant{ id }
+          channel{ slug }
+        }
+      }
+    }
+"""
+
+PRODUCT_VARIANT_BACK_IN_STOCK_FOR_CLICK_AND_COLLECT = """
+    subscription{
+      event{
+        ...on ProductVariantBackInStockForClickAndCollect{
+          productVariant{ id }
+          channel{ slug }
+        }
+      }
+    }
+"""
+
+
+@pytest.fixture
+def subscription_product_variant_out_of_stock_in_channel_webhook(subscription_webhook):
+    return subscription_webhook(
+        PRODUCT_VARIANT_OUT_OF_STOCK_IN_CHANNEL,
+        WebhookEventAsyncType.PRODUCT_VARIANT_OUT_OF_STOCK_IN_CHANNEL,
+    )
+
+
+@pytest.fixture
+def subscription_product_variant_back_in_stock_in_channel_webhook(subscription_webhook):
+    return subscription_webhook(
+        PRODUCT_VARIANT_BACK_IN_STOCK_IN_CHANNEL,
+        WebhookEventAsyncType.PRODUCT_VARIANT_BACK_IN_STOCK_IN_CHANNEL,
+    )
+
+
+@pytest.fixture
+def subscription_product_variant_out_of_stock_for_click_and_collect_webhook(
+    subscription_webhook,
+):
+    return subscription_webhook(
+        PRODUCT_VARIANT_OUT_OF_STOCK_FOR_CLICK_AND_COLLECT,
+        WebhookEventAsyncType.PRODUCT_VARIANT_OUT_OF_STOCK_FOR_CLICK_AND_COLLECT,
+    )
+
+
+@pytest.fixture
+def subscription_product_variant_back_in_stock_for_click_and_collect_webhook(
+    subscription_webhook,
+):
+    return subscription_webhook(
+        PRODUCT_VARIANT_BACK_IN_STOCK_FOR_CLICK_AND_COLLECT,
+        WebhookEventAsyncType.PRODUCT_VARIANT_BACK_IN_STOCK_FOR_CLICK_AND_COLLECT,
+    )
 
 
 def test_product_variant_out_of_stock_in_channel(
